+++ src/index.ts
@@ -7,9 +7,13 @@ import {
 } from './accounts'
 import {
   addOwner,
+  changeThreshold,
+  disableEcdsa,
+  disablePasskeys,
+  enableEcdsa,
+  enablePasskeys,
   recover,
   removeOwner,
-  setThreshold,
   setUpRecovery,
 } from './actions'
 import type { TransactionResult } from './execution'
@@ -172,9 +176,13 @@ async function createRhinestoneAccount(
 export {
   createRhinestoneAccount,
   addOwner,
+  changeThreshold,
+  disableEcdsa,
+  disablePasskeys,
+  enableEcdsa,
+  enablePasskeys,
   recover,
   removeOwner,
-  setThreshold,
   setUpRecovery,
 }
 export type {