+++ src/modules/validators/core.ts
@@ -21,7 +21,6 @@ interface PublicKey {
 interface WebauthnCredential {
   pubKey: PublicKey | Hex | Uint8Array
   authenticatorId: string
-  hook?: Address
 }
 
 const OWNABLE_VALIDATOR_ADDRESS: Address =
@@ -186,7 +185,10 @@ function parsePublicKey(publicKey: Hex | Uint8Array): PublicKey {
 export {
   OWNABLE_VALIDATOR_ADDRESS,
   getOwnerValidator,
+  getOwnableValidator,
+  getWebAuthnValidator,
   getSocialRecoveryValidator,
   getValidator,
   getMockSignature,
 }
+export type { WebauthnCredential }